Replace outdated documentation with a link to user documentation
While users should always refer to the user documentation as the sole
source of truth, some might still refer to this blog post. To prevent
misunderstands and issues, we point them to the user documentation.

Updating old blog posts whenever a step changes doesn't scale and it's
easy to forget.